jack_delay.cc will need to be hand-ported
jack_delay.cc is removed because it contains too many dependnecies that will make it difficult to compile via emuscripten. However mtdm is the core algorithm which should more easily be compiled with emuscripten. The essential code in jack_delay.cc is the while (active) loop, which can easily be hand-ported to javascript, while the majority of the rest of the code is jack-specific binding code which can't be used in the browser anyway.